pyproject.toml
[project]
name = "hr-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.21",
    "faiss-cpu>=1.12.0",
    "jsonpointer>=3.0.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "langgraph-cli[inmem]>=0.4.2",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
]
